Initial refactor of server_config and filetype parsers
move wilfred/parser to wilfred/api/parser (update imports)
move wilfred/server_config to wilfred/api/server_config
improve CLI exception handling (add spinner.fail() before ui_exception is called on Halo spinners)
